% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiPlazas.R
\name{apiPlazas}
\alias{apiPlazas}
\title{Devuelve los datos de plazas libres de los parkings del polígono}
\usage{
apiPlazas()
}
\value{
json
}
\description{
Devuelve los datos de plazas libres de los parkings del polígono
}
\examples{
 apiPlazas()

}
